Undo automatic escaping of % signs
It was a bad idea to automatically escape % signs. This causes all sorts
of issues (e.g. if %-sign is used in a string which is intended for the
underlying app) and is based on the assumption that the YML author does
not know about the special meaning of %-signs.

For now, we will just stop escaping automatically. This might lead to
weird diffs (which was the original motivation to have this), but at
least it is safe and does not cause side-effects.

Closes #4.