Rewrite diff engine
Using apply was not fully correct (e.g. fields cannot be deleted that
way) and potentially dangerois (as the template overwrites the platform
version).

Using JSON patches is fully correct and minimal (only paths needing to
change are touched). The downside of it is that it is way more complex
to deal with ...

Closes #26.