Use "oc apply" instead of generating patches
This makes use of `oc apply` (client-side apply), as this is more solid
than the custom generation of patches used with `oc patch`.

Also, new resources are created with `oc apply` instead of `oc create`,
which has the advantage that Tailor does not fail when the desired state
defines a resource with a label that is not present in the current
state. Previously, Tailor did not "see" the resource in the current
state because it was not in scope, and therefore tried to create if
which failed. Now Tailor creates with `apply`, which ends up being a
merge, which is better.

The change removes the custom Tailor annotations, it just makes use of
the `kubectl.kubernetes.io/last-applied-configuration` annotation.

This annotation is used during client-side apply, which is also known to
cause problems, which is why Kubernetes has introduced a server-side
apply, which is based on field ownership. Long-term, this is something
that Tailor could use as well, but right now, Tailor targets OpenShift
v3, which ses a Kubernetes version which only supports client-side
apply, and therefore using the annotation is a good fit.

If there are any issues with unexpected drift or problems when
reconciling, please consult
https://kubernetes.io/docs/tasks/manage-kubernetes-objects/declarative-config/.